+++ demisto_sdk/commands/common/hook_validations/id.py
@@ -1,4 +1,3 @@
-import json
 import os
 import re
 from collections import OrderedDict
@@ -18,9 +17,8 @@ from demisto_sdk.commands.common.update_id_set import (get_classifier_data,
 from demisto_sdk.commands.unify.unifier import Unifier
 
 
-class IDSetValidator(BaseValidator):
-    """IDSetValidator was designed to validate the inter connectivity between content entities
-     in the content, private repo and premium repo.
+class IDSetValidations(BaseValidator):
+    """IDSetValidations was designed to make sure all the inter connected content entities are valid.
 
     The id_set.json file is created using the update_id_set.py script. It contains all the data from the various
     executables we have in Content repository -
@@ -43,41 +41,22 @@ class IDSetValidator(BaseValidator):
     MAPPERS_SECTION = "Mappers"
     INCIDENT_TYPES_SECTION = "IncidentTypes"
 
-    ID_SET_PATH = "./Tests/id_set.json"
-
     def __init__(self, is_test_run=False, is_circle=False, configuration=Configuration(), ignored_errors=None,
-                 print_as_warnings=False, suppress_print=False):
+                 print_as_warnings=False, suppress_print=False, id_set_file=None):
         super().__init__(ignored_errors=ignored_errors, print_as_warnings=print_as_warnings,
                          suppress_print=suppress_print)
         self.is_circle = is_circle
         self.configuration = configuration
         if not is_test_run and self.is_circle:
-            self.id_set = self.load_id_set()
-            self.id_set_path = os.path.join(self.configuration.env_dir, 'configs', 'id_set.json')
-            self.script_set = self.id_set[self.SCRIPTS_SECTION]
-            self.playbook_set = self.id_set[self.PLAYBOOK_SECTION]
-            self.integration_set = self.id_set[self.INTEGRATION_SECTION]
-            self.test_playbook_set = self.id_set[self.TEST_PLAYBOOK_SECTION]
-            self.classifiers_set = self.id_set[self.CLASSIFIERS_SECTION]
-            self.layouts_set = self.id_set[self.LAYOUTS_SECTION]
-            self.mappers_set = self.id_set[self.MAPPERS_SECTION]
-            self.incident_types_set = self.id_set[self.INCIDENT_TYPES_SECTION]
-
-    def load_id_set(self):
-        with open(self.ID_SET_PATH, 'r') as id_set_file:
-            try:
-                id_set = json.load(id_set_file)
-            except ValueError as ex:
-                if "Expecting property name" in str(ex):
-                    error_message, error_code = Errors.id_set_conflicts()
-                    if self.handle_error(error_message, error_code, file_path="id_set.json"):
-                        raise
-                    else:
-                        pass
-
-                raise
-
-            return id_set
+            self.id_set_file = id_set_file
+            self.script_set = self.id_set_file[self.SCRIPTS_SECTION]
+            self.playbook_set = self.id_set_file[self.PLAYBOOK_SECTION]
+            self.integration_set = self.id_set_file[self.INTEGRATION_SECTION]
+            self.test_playbook_set = self.id_set_file[self.TEST_PLAYBOOK_SECTION]
+            self.classifiers_set = self.id_set_file[self.CLASSIFIERS_SECTION]
+            self.layouts_set = self.id_set_file[self.LAYOUTS_SECTION]
+            self.mappers_set = self.id_set_file[self.MAPPERS_SECTION]
+            self.incident_types_set = self.id_set_file[self.INCIDENT_TYPES_SECTION]
 
     def _is_incident_type_default_playbook_found(self, incident_type_data):
         """Check if the default playbook of an incident type is in the id_set
@@ -267,6 +246,7 @@ class IDSetValidator(BaseValidator):
             bool. Whether the file is valid in the id_set or not.
         """
         is_valid = True
+
         if self.is_circle:  # No need to check on local env because the id_set will contain this info after the commit
             click.echo(f"id set validations for: {file_path}")