+++ demisto_sdk/commands/find_dependencies/find_dependencies.py
@@ -9,6 +9,7 @@ from typing import Union
 import click
 import networkx as nx
 from demisto_sdk.commands.common import constants
+from demisto_sdk.commands.common.constants import GENERIC_COMMANDS_NAMES
 from demisto_sdk.commands.common.tools import print_error
 from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
 
@@ -324,7 +325,9 @@ class PackDependencies:
             command_to_integration = list(script.get('command_to_integration', {}).keys())
             script_executions = script.get('script_executions', [])
 
-            dependencies_commands = list(set(depends_on + command_to_integration + script_executions))
+            all_dependencies_commands = list(set(depends_on + command_to_integration + script_executions))
+            dependencies_commands = list(filter(lambda cmd: cmd not in GENERIC_COMMANDS_NAMES,
+                                                all_dependencies_commands))  # filter out generic commands
 
             for command in dependencies_commands:
                 # try to search dependency by scripts first
@@ -417,10 +420,11 @@ class PackDependencies:
             implementing_commands_and_integrations = playbook_data.get('command_to_integration', {})
 
             for command, integration_name in implementing_commands_and_integrations.items():
+                packs_found_from_integration = set()
                 if integration_name:
                     packs_found_from_integration = PackDependencies._search_packs_by_items_names(
                         integration_name, id_set['integrations'], exclude_ignored_dependencies)
-                else:
+                elif command not in GENERIC_COMMANDS_NAMES:  # do not collect deps on generic command in Pbs
                     packs_found_from_integration = PackDependencies._search_packs_by_integration_command(
                         command, id_set, exclude_ignored_dependencies)
 
@@ -568,7 +572,7 @@ class PackDependencies:
             #    The CommonTypes pack is not dependent on the Zimperium Pack, but vice versa.
             # 2. emailfrom in the Phishing pack is being used in the EWS pack.
             #    Phishing pack does not depend on EWS but vice versa.
-            # The opposite dependencies are calculated in: _collect_playbook_dependencies, _collect_mapper_dependencies
+            # The opposite dependencies are calculated in: _collect_playbook_dependencies, _collect_mappers_dependencies
 
             related_scripts = incident_field_data.get('scripts', [])
             packs_found_from_scripts = PackDependencies._search_packs_by_items_names(